Extract default home page into separate component
